5/18 pass

+++ digit/lib/end_to_end_tests/stellar_soroban/unofficial_digicus_examples/answer_to_life/src/lib.rs
@@ -1,12 +1,13 @@
 #![no_std]
-use soroban_sdk::{contract, contractimpl, symbol_short, vec, Env, Symbol, Vec, log};
+use soroban_sdk::{contract, contractimpl, Env};
 
 #[contract]
 pub struct AnswerToLifeContract;
 
+
 #[contractimpl]
 impl AnswerToLifeContract {
-    pub fn fourty_two(env: Env) -> u32 {
+    pub fn fourty_two(env: Env) -> i64 {
         42
     }
 }
